add opacities package
