+++ benchmarks/transport_montecarlo_packet.py
@@ -5,14 +5,13 @@ Basic TARDIS Benchmark.
 import numpy as np
 from asv_runner.benchmarks.mark import parameterize, skip_benchmark
 
-import tardis.transport.montecarlo.estimators.radfield_mc_estimators
-import tardis.transport.montecarlo.estimators.radfield_mc_estimators
-import tardis.transport.montecarlo.numba_interface as numba_interface
-import tardis.transport.montecarlo.opacities as opacities
-import tardis.transport.montecarlo.r_packet as r_packet
-import tardis.transport.montecarlo.utils as utils
+import tardis.opacities.opacities as opacities
 import tardis.transport.frame_transformations as frame_transformations
 import tardis.transport.geometry.calculate_distances as calculate_distances
+import tardis.transport.montecarlo.estimators.radfield_mc_estimators
+import tardis.transport.montecarlo.numba_interface as numba_interface
+import tardis.transport.montecarlo.r_packet as r_packet
+import tardis.transport.montecarlo.utils as utils
 import tardis.transport.r_packet_transport as r_packet_transport
 from benchmarks.benchmark_base import BenchmarkBase
 from tardis.model.geometry.radial1d import NumbaRadial1DGeometry
