+++ tardis/energy_input/gamma_ray_transport.py
@@ -1,14 +1,15 @@
 import logging
+
+import astropy.units as u
 import numpy as np
 import pandas as pd
-import astropy.units as u
 import radioactivedecay as rd
 
 from tardis.energy_input.energy_source import (
     get_all_isotopes,
     setup_input_energy,
 )
-from tardis.transport.montecarlo.opacities import M_P
+from tardis.opacities.opacities import M_P
 
 # Energy: keV, exported as eV for SF solver
 # distance: cm
@@ -317,10 +318,12 @@ def decay_chain_energies(
 
 def fractional_decay_energy(decay_energy):
     """Function to calculate fractional decay energy
+
     Parameters
     ----------
     decay_energy : Dict
         dictionary of decay chain energies for each isotope in each shell
+
     Returns
     -------
     fractional_decay_energy : Dict
