+++ tardis/plasma/properties/radiative_properties.py
@@ -16,7 +16,6 @@ logger = logging.getLogger(__name__)
 
 __all__ = [
     "StimulatedEmissionFactor",
-    "TauSobolev",
     "BetaSobolev",
     "TransitionProbabilities",
     "RawRadBoundBoundTransProbs",
@@ -113,75 +112,6 @@ class StimulatedEmissionFactor(ProcessingPlasmaProperty):
                 (stimulated_emission_factor < 0) & nlte_lines_mask[np.newaxis].T
             ] = 0.0
         return stimulated_emission_factor
-
-
-class TauSobolev(ProcessingPlasmaProperty):
-    """
-    Attributes
-    ----------
-    tau_sobolev : Pandas DataFrame, dtype float
-          Sobolev optical depth for each line. Indexed by line.
-          Columns as zones.
-    """
-
-    outputs = ("tau_sobolevs",)
-    latex_name = (r"\tau_{\textrm{sobolev}}",)
-    latex_formula = (
-        r"\dfrac{\pi e^{2}}{m_{e} c}f_{lu}\lambda t_{exp}\
-        n_{lower} \Big(1-\dfrac{g_{lower}n_{upper}}{g_{upper}n_{lower}}\Big)",
-    )
-
-    def __init__(self, plasma_parent):
-        super(TauSobolev, self).__init__(plasma_parent)
-        self.sobolev_coefficient = (
-            (
-                ((np.pi * const.e.gauss**2) / (const.m_e.cgs * const.c.cgs))
-                * u.cm
-                * u.s
-                / u.cm**3
-            )
-            .to(1)
-            .value
-        )
-
-    def calculate(
-        self,
-        lines,
-        level_number_density,
-        lines_lower_level_index,
-        time_explosion,
-        stimulated_emission_factor,
-        j_blues,
-        f_lu,
-        wavelength_cm,
-    ):
-        f_lu = f_lu.values[np.newaxis].T
-        wavelength = wavelength_cm.values[np.newaxis].T
-        n_lower = level_number_density.values.take(
-            lines_lower_level_index, axis=0, mode="raise"
-        )
-        tau_sobolevs = (
-            self.sobolev_coefficient
-            * f_lu
-            * wavelength
-            * time_explosion
-            * n_lower
-            * stimulated_emission_factor
-        )
-
-        if np.any(np.isnan(tau_sobolevs)) or np.any(
-            np.isinf(np.abs(tau_sobolevs))
-        ):
-            raise ValueError(
-                "Some tau_sobolevs are nan, inf, -inf in tau_sobolevs."
-                " Something went wrong!"
-            )
-
-        return pd.DataFrame(
-            tau_sobolevs,
-            index=lines.index,
-            columns=np.array(level_number_density.columns),
-        )
 
 
 class BetaSobolev(ProcessingPlasmaProperty):
