+++ tardis/transport/montecarlo/tests/test_packet.py
@@ -1,14 +1,14 @@
 import numpy as np
 import pytest
 
-import tardis.transport.montecarlo.montecarlo_configuration as numba_config
-import tardis.transport.montecarlo.estimators.radfield_mc_estimators
-import tardis.transport.montecarlo.numba_interface as numba_interface
-import tardis.transport.montecarlo.opacities as opacities
-import tardis.transport.montecarlo.r_packet as r_packet
-import tardis.transport.montecarlo.utils as utils
+import tardis.opacities.opacities as opacities
 import tardis.transport.frame_transformations as frame_transformations
 import tardis.transport.geometry.calculate_distances as calculate_distances
+import tardis.transport.montecarlo.estimators.radfield_mc_estimators
+import tardis.transport.montecarlo.montecarlo_configuration as numba_config
+import tardis.transport.montecarlo.numba_interface as numba_interface
+import tardis.transport.montecarlo.r_packet as r_packet
+import tardis.transport.montecarlo.utils as utils
 import tardis.transport.r_packet_transport as r_packet_transport
 from tardis import constants as const
 from tardis.model.geometry.radial1d import NumbaRadial1DGeometry
